Move getErrorsInFormat (draft) from the datatypes file to draftutils.
